+++ lib/ts/recipe/passwordless/components/themes/signInUp/phoneNumberInput.tsx
@@ -16,20 +16,18 @@
 /*
  * Imports.
  */
+// import phoneNumberInputLibStyles from "react-phone-number-input/style.css";
+import phoneNumberInputLibStyles from "intl-tel-input/build/css/intlTelInput.css";
+import intlTelInput from "intl-tel-input";
 
-import { CSSObject } from "@emotion/react";
-import Select, { components as ReactSelectComps } from "react-select";
-
-import React, { useCallback, useEffect, useMemo } from "react";
-
-import PhoneInputWithCountrySelect, { getCountryCallingCode, getCountries } from "react-phone-number-input/min";
+// eslint-disable no-unused-vars
+import React, { useCallback, useEffect, useRef } from "react";
 
 import { InputProps } from "../../../../emailpassword/components/library/input";
 import ErrorIcon from "../../../../../components/assets/errorIcon";
-import { CountryCode } from "libphonenumber-js";
 
 type PhoneNumberInputProps = {
-    defaultCountry?: CountryCode;
+    defaultCountry?: string;
 };
 
 /*
@@ -44,7 +42,6 @@ function PhoneNumberInput({
     onInputFocus,
     onChange,
     hasError,
-    placeholder,
     value,
 }: InputProps & PhoneNumberInputProps): JSX.Element {
     function handleFocus() {
@@ -65,19 +62,46 @@ function PhoneNumberInput({
         }
     }
 
-    function handleChange(newValue: string) {
-        if (onChange !== undefined) {
-            onChange({
-                id: name,
-                value: newValue,
-            });
-        }
-    }
+    const onChangeRef = useRef(onChange);
+    useEffect(() => {
+        onChangeRef.current = onChange;
+    }, [onChange]);
+
+    const handleChange = useCallback(
+        (newValue: string) => {
+            if (onChangeRef.current !== undefined) {
+                onChangeRef.current({
+                    id: name,
+                    value: newValue,
+                });
+            }
+        },
+        [onChangeRef]
+    );
+
+    const handleCountryChange = useCallback(
+        (ev) => {
+            if (onChangeRef.current !== undefined) {
+                onChangeRef.current({
+                    id: name,
+                    value: ev.target.value,
+                });
+            }
+        },
+        [onChangeRef]
+    );
+
+    const inputRef = useRef<HTMLInputElement | null>(null);
 
     useEffect(() => {
-        const countries = getCountries();
-        for (const code of countries) {
-            new Image().src = `http://purecatamphetamine.github.io/country-flag-icons/3x2/${code}.svg`;
+        if (inputRef.current !== null && inputRef.current.dataset["intl-tel-input-id"] === undefined) {
+            inputRef.current.value = value;
+            intlTelInput(inputRef.current, {
+                initialCountry: defaultCountry,
+                nationalMode: false,
+                preferredCountries: defaultCountry ? [defaultCountry] : [],
+            });
+            inputRef.current.addEventListener("countrychange", handleCountryChange);
         }
     }, []);
 
@@ -86,28 +110,32 @@ function PhoneNumberInput({
      */
     return (
         <div data-supertokens="inputContainer">
-            <div data-supertokens={`inputWrapper ${hasError ? "inputError" : ""}`}>
-                <PhoneInputWithCountrySelect
-                    data-supertokens="input phoneInputLibRoot"
-                    countrySelectComponent={CountrySelectWithIcon}
+            <style type="text/css">
+                {
+                    // There should be a better way around this... :/
+                    phoneNumberInputLibStyles.replace(":root", '[data-supertokens~="container"]')
+                }
+                {`
+                    .iti__flag {background-image: url("https://cdnjs.cloudflare.com/ajax/libs/intl-tel-input/17.0.19/img/flags.png");}
+
+                    @media (-webkit-min-device-pixel-ratio: 2), (min-resolution: 192dpi) {
+                        .iti__flag {background-image: url("https://cdnjs.cloudflare.com/ajax/libs/intl-tel-input/17.0.19/img/flags@2x.png");}
+                    }
+                `}
+            </style>
+            <div data-supertokens={`phoneInputWrapper inputWrapper ${hasError ? "inputError" : ""}`}>
+                <input
+                    type="tel"
+                    data-supertokens="input"
                     name={name + "_text"}
                     autoFocus={autofocus}
                     autoComplete={autoComplete}
-                    value={value}
-                    onChange={handleChange}
-                    countryCallingCodeEditable={true}
+                    onChange={(ev) => {
+                        handleChange(ev.target.value);
+                    }}
                     onFocus={handleFocus}
                     onBlur={handleBlur}
-                    international={true}
-                    placeholder={placeholder}
-                    defaultCountry={defaultCountry}
-                    smartCaret={
-                        false
-                        /*
-                            This is set to false to avoid this bug: https://gitlab.com/catamphetamine/react-phone-number-input/-/issues/128
-                            We should monitor this and update it to true whenever it's fixed
-                        */
-                    }
+                    ref={inputRef}
                 />
                 {hasError === true && (
                     <div data-supertokens="inputAdornment inputAdornmentError">
@@ -116,101 +144,6 @@ function PhoneNumberInput({
                 )}
             </div>
         </div>
-    );
-}
-
-function CountrySelectWithIcon({
-    value,
-    options,
-    iconComponent: Icon,
-    onChange,
-}: {
-    value: string;
-    options: { value: string; label: string; divider: boolean }[];
-    iconComponent: React.ComponentClass<any>;
-    onChange: (newValue: string | undefined) => void;
-}) {
-    const style: any = {};
-
-    const selectedOption = options.find((o) => o.value === value);
-
-    const selectStyles = useMemo(
-        () => ({
-            menu: (provided: CSSObject) => ({
-                ...provided,
-                ...style.phoneInputCountryDropdown,
-                // These elements are added inside CustomOption below
-                // We are styling them here, to avoid emotion processing on each element (200+ countries)
-                "& [data-supertokens=phoneInputCountryOptionLabel]": {
-                    ...style.phoneInputCountryOptionLabel,
-                },
-                "& [data-supertokens=phoneInputCountryOption]": {
-                    ...style.phoneInputCountryOption,
-                },
-                "& [data-supertokens=phoneInputCountryOptionCallingCode]": {
-                    ...style.phoneInputCountryOptionCallingCode,
-                },
-            }),
-            control: (provided: CSSObject) => ({
-                ...provided,
-                ...style.phoneInputCountryControl,
-            }),
-            valueContainer: (provided: CSSObject) => ({
-                ...provided,
-                ...style.phoneInputCountryValueContainer,
-            }),
-            dropdownIndicator: (provided: CSSObject) => ({
-                ...provided,
-                ...style.phoneInputCountryDropdownIndicator,
-            }),
-        }),
-        [style]
-    );
-    const selectOnChange = useCallback(
-        (selected) => {
-            onChange(selected === null ? undefined : selected.value);
-        },
-        [onChange]
-    );
-    const components = useMemo(() => {
-        const CustomOption = ({ innerProps, data, isSelected, isDisabled }: any) => {
-            return !isDisabled ? (
-                <div
-                    {...innerProps}
-                    data-supertokens="phoneInputCountryOption"
-                    onTouchEnd={(ev) => {
-                        // We need to stop propagation here, to prevent the menu closing before the click is fired
-                        ev.stopPropagation();
-                    }}
-                    aria-selected={isSelected}>
-                    <Icon country={data.value} label={data.label} />
-                    <span data-supertokens="phoneInputCountryOptionLabel">{data.label}</span>
-                    {data.value && (
-                        <span data-supertokens="phoneInputCountryOptionCallingCode">
-                            +{getCountryCallingCode(data.value)}
-                        </span>
-                    )}
-                </div>
-            ) : null;
-        };
-        const Control = ({ children, ...rest }: any) => {
-            return (
-                <ReactSelectComps.SingleValue data-supertokens="phoneInputCountrySelect" {...rest}>
-                    <Icon country={rest.getValue()[0]?.value} label={children} />
-                </ReactSelectComps.SingleValue>
-            );
-        };
-        return { Option: CustomOption, SingleValue: Control, IndicatorSeparator: undefined };
-    }, []);
-
-    return (
-        <Select
-            options={options}
-            styles={selectStyles}
-            value={selectedOption}
-            onChange={selectOnChange}
-            components={components}
-        />
     );
 }